Split up the settings class in multiple classes
