+++ src/main/java/settings/bubbleSettings.java
@@ -1,0 +1,39 @@
+package settings;
+
+public class bubbleSettings {
+
+    private static int bubbleDefaultDiameter = 16;
+    private static int bubbleDefaultX = 10;
+    private static int bubbleDefaultY = 10;
+
+    private static int smallestBubbleSize = 5;
+
+    /**
+     * @return the bubbleDefaultDiameter
+     */
+    public static int getBubbleDefaultDiameter() {
+        return bubbleDefaultDiameter;
+    }
+
+    /**
+     * @return the bubbleDefaultX
+     */
+    public static int getBubbleDefaultX() {
+        return bubbleDefaultX;
+    }
+
+    /**
+     * @return the bubbleDefaultY
+     */
+    public static int getBubbleDefaultY() {
+        return bubbleDefaultY;
+    }
+
+    /**
+     * @return the smallestBubbleSize
+     */
+    public static int getSmallestBubbleSize() {
+        return smallestBubbleSize;
+    }
+
+}
