+++ src/main/java/settings/ropeSettings.java
@@ -1,0 +1,14 @@
+package settings;
+
+public class ropeSettings {
+
+    private static int ropeSpeed = 4;
+
+    /**
+     * @return the ropeSpeed
+     */
+    public static int getRopeSpeed() {
+        return ropeSpeed;
+    }
+
+}
